format code to clang format

+++ march_hardware/include/march_hardware/IMotionCubeState.h
@@ -95,7 +95,7 @@ public:
   std::string motionErrorDescription;
   float motorCurrent;
   float motorVoltage;
-  float incrementalEncoderValue;
+  int incrementalEncoderValue;
 };
 
 }  // namespace march
